requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
Django==5.0.4
django-allauth==0.62.1
django-axes==6.4.0
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-environ==0.11.2
django-ipware==7.0.1
django-preventconcurrentlogins==0.8.2
django-ratelimit==4.1.0
django-tailwind==3.8.0
et-xmlfile==1.1.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.22
Pygments==2.17.2
pypng==0.20220715.0
pytest==8.2.2
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
ratelimit==2.2.1
reportlab==4.2.0
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
